% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_estimation}
\alias{atlas_estimation}
\title{Atlas with estimation of number of cells per brain area by Erö et al.}
\format{
A data frame with 955 rows and 10 variables:
\describe{
  \item{Regions}{name of the brain areas, consistent with ABA (25µm, mouse), corresponds to name var in data("atlas_tree")}
  \item{Cells}{number of total cells estimated in each brain area}
  \item{Neurons}{number of neurons estimated in each brain area}
  \item{Glia}{number of glia estimated in each brain area}
  \item{Excitatory}{number of excitatory cells estimated in each brain area}
  \item{Inhibitory}{number of inhibitory cells estimated in each brain area}
  \item{Modulatory}{number of modulatory cells estimated in each brain area}
  \item{Astrocytes}{number of astrocytes estimated in each brain area}
  \item{Olygodendrocytes}{number of olygodendrocytes estimated in each brain area}
  \item{Microglia}{number of microglia estimated in each brain area}
}
}
\source{
\url{https://doi.org/10.3389/fninf.2018.00084}
}
\usage{
atlas_estimation
}
\description{
A dataset published by Erö and colleagues (doi: 10.3389/fninf.2018.00084) containing an estimation of number
of cells per brain areas in Allen Brain Reference atlas (25µm, mouse) space.
}
\keyword{datasets}
