% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_density.R
\name{calculate_density}
\alias{calculate_density}
\title{Calculate density of active cells}
\usage{
calculate_density(cells_df, n_slice = 10, path)
}
\arguments{
\item{cells_df}{dataframe with xyz coordinates ("xPos", "yPos", "zPos") of the brain area of interest.
Each row corresponds to a different cell that belongs to a sample ("sample_id")}

\item{n_slice}{integer which specifies in how many parts to slice the estimate. Default = 10.}

\item{path}{specify path were data for each sample separtely will be saved.}
}
\value{

}
\description{
Wrapper around kde() function from the ks package to calculate the density of
a brain area across samples. It saves each sample separately in a specified folder.
}
\examples{
x <- data.frame(
sample_id = rep(c(1:3), each = 100),
xPos = c(rnorm(100), rpois(100, 2), sample(100),
yPos = c(rnorm(100), rpois(100, 2), sample(100),
zPos = c(rnorm(100), rpois(100, 2), sample(100)
)
}
