% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{cleaned_counts_to_df}
\alias{cleaned_counts_to_df}
\title{Merges in one data frame all cleaned annotated cells}
\usage{
cleaned_counts_to_df(path_cleaned)
}
\arguments{
\item{path_cleaned}{String to specify path where cleaned annotated cells will be saved}
}
\value{

}
\description{
This function is a wrapper to dplyr functions to merge together
the multiple files output from cleaned_counts(). Only RDS files with extension
'_clean_cells.RDS' will be used.
}
\examples{
For a thorough example, please see XX.
}
