% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Miniature example of whole-brain imaging data over time from Bonapersona et al.}
\format{
A data frame in long format with each row indicating a c-fos+ cells. Contains 177885 rows and 7 variables:
\describe{
  \item{sample_id}{unique identified for each sample}
  \item{xPos}{position on the x-axis in horizontal sectioning, consistent with ABA (25µm, mouse)}
  \item{yPos}{position on the y-axis in horizontal sectioning, consistent with ABA (25µm, mouse)}
  \item{zPos}{position on the z-axis in horizontal sectioning, consistent with ABA (25µm, mouse)}
  \item{maxInt}{maximum expression level of c-fos measured per cell, estimated with Imaris}
  \item{my_grouping}{grouping of the brain areas based on the researchers specified spatial resolution, consistent with acronyms from ABA (25µm, mouse)}
  \item{hemisphere}{hemisphere}
}
}
\source{
\url{https://osf.io/8muvw/}
}
\usage{
example
}
\description{
A dataset containing the cleaned coordinates of c-fos positive cells in a selection of brain areas
(amygdala and hippocampus) after foot-shock. Experimental groups are: 30, 90, 180 minutes after
foot-shock and a sham control (0 min). For more information, see https://osf.io/8muvw/ .
}
\keyword{datasets}
