% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_brain_areas.R
\name{find_intercept}
\alias{find_intercept}
\title{Function to find where an interpolated linear curve crosses a y value}
\usage{
find_intercept(x, y, inter = 0)
}
\arguments{
\item{x}{number of x variable}

\item{y}{number of y variable}

\item{inter}{number of y at which you want to calculate x}
}
\value{

}
\description{
The function interpolates a linear model within two points, and finds at
which x the curve has a specified y.
}
\examples{
find_intercept(x = c(-1.1, 1.1) , y = c(-1.1, 0.9))
}
