% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_aba_atlas.R
\name{find_parents}
\alias{find_parents}
\title{Find larger grouping areas (parents) of a smaller brain area
(child) in the Allen Brain Atlas.}
\usage{
find_parents(atlas, child)
}
\arguments{
\item{atlas}{Data frame with (at least) two character variables: one for the parents (parent_acronym),
and one for the children (acronym).}

\item{child}{Acronym of the child brain area. It must be a character vector of
length 1.}
}
\value{

}
\description{
The Allen Brain Atlas is organized with a tree structure, where
 smaller brain areas (e.g. nuclei) are grouped into larger brain areas. This function
 allows you to find the parent of a sub-brain area of choice. Returns only one level up
 in the tree structure.
}
\examples{
#' x <- data.frame(
parent_acronym = c(rep("main_parent", 3), rep("child_with_children", 2), "grandchild_with_children"),
acronym = c("main_parent", "child_with_children", "child_without_children", "grandchild_with_children", "grandchild_without_children", "great_grandchild"))
this_child <- "child_with_children"

find_parents(atlas = x, child = this_child)
}
