% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_most_active.R
\name{from_aba_api_to_df}
\alias{from_aba_api_to_df}
\title{Convert data downloaded from Allen Brain Atlas API to a dataframe}
\usage{
from_aba_api_to_df(api_in_csv)
}
\arguments{
\item{api_in_csv}{data downloaded from Allen Brain Atlas api in .csv format}
}
\value{

}
\description{
The data must have been previously downloaded from the Allen Brain Atlas API, and saved as .csv.
The function converts the data so that the final output is a dataframe with a row for each brain area ("acronym"),
its full name ("name") and the mean ("mean_expression") as well as standard deviation ("sd_expression") of the protein of
interest across all Allen Brain Atlas successful experiments.
}
\examples{
FOR EXAMPLE SEE X.
}
