% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mask}
\alias{mask}
\title{Mask for data cleaning of whole-brain microscopy.}
\format{
A data frame with 1205 rows and 5 variables:
\describe{
  \item{edge}{locations around edges of the brain, size of 75µm, consistent with ABA (25µm, mouse)}
  \item{edge}{locations around the ventricles, size of 75µm, consistent with ABA (25µm, mouse)}
}
}
\source{
\url{add link to git}
}
\usage{
mask
}
\description{
A list containing two datasets to remove the edges of the brain (dataframe edge) and the space
around the ventricles (dataframe ventricles) for cleaning whole-brain microscopy images. The
datasets are on the Allen Brain Reference Atlas (25µm, mouse) brain. The abc4d package uses functions to
remove these coordinates from the samples.
}
\keyword{datasets}
