% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{normalize}
\alias{normalize}
\title{Normalize cell counts}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric vector with values to normalize}
}
\value{

}
\description{
Wrapper functon to YeoJohnson transformation from recipes package. Similar transformation to boxcox,
but it accepts also 0 (and negative) values.
}
\examples{
x <- rpois(100, 3)
y <- normalize(x)

par(mfrow=c(1,2))
hist(x)
hist(y)
par(mfrow=c(1,1))
}
