% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_brain_areas.R
\name{order_brain_areas}
\alias{order_brain_areas}
\title{Order time of activation of brain areas based on their expression levels}
\usage{
order_brain_areas(
  region_df,
  order_of_groups,
  count_var = "cells_perthousand_box_scaled_ba"
)
}
\arguments{
\item{region_df}{region_based dataframe. Each row is a brain area ("my_grouping") per sample, with a variable for counts
(e.g. "cells_perthousand_box_scaled_ba"). The variable "group" specifies the experimental group, and it is used for the ordering.
It can be output from summarize_per_region() or preprocess_per_region().}

\item{order_of_groups}{string with the same unique elements as region_df$group, but ordered as should be used for the analysis.}

\item{count_var}{variable specified by the user on which to conduct the analysis.}
}
\value{

}
\description{
By using expression levels in different groups (e.g. time points),
brain areas are ordered .
}
\examples{
# baby example
x <- data.frame(
batch = rep(c("1", "2"), 2),
group = rep(rep(c("0", "90"), each = 2), 2),
my_grouping = "CA1",
cells_perthousand_box_scaled_ba = c(-2, -1, 2, 1, -3, -2, 4, 2)
)
order_brain_areas(x, order_of_groups = c("0", "90"))

x <- data.frame(
batch = rep(c(1,1,2,2), each = 5),
group = rep(c("control", "exp", "exp", "control"), each = 5),
sample_id = rep(c("a", "b", "c", "d"), each = 5),
my_grouping = rep(c("CA1", "CA2", "CA3", "DG", "BLA"), 4),
intensity_ave = sample(10000, 20, replace = TRUE),
cells_perthousand = abs(rnorm(20))
)

find_most_active(x)
}
