% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_analysis.R
\name{pairwise_activation}
\alias{pairwise_activation}
\title{Calculate pairwise activation across groups}
\usage{
pairwise_activation(region_df, against_group = "0", heatmap = FALSE)
}
\arguments{
\item{region_df}{region_based dataframe. It can be the output of preprocess_per_region(). Each row corresponds
to a brain area ("my_grouping") per sample, with (corrected) counts ("cells_perthousand_box_scaled"). The variable "group" refers to the experimental groups. The variable
"batch" is used for pairwise comparisons. Of note, for each "batch" there can be only a sample per group.}

\item{against_group}{group from "group" var against which comparisons will be made}

\item{heatmap}{if TRUE, the function returns a list where the first element is the dataframe, and the second
element is a visualization with heatmap.}

\item{count_var}{string with the name of the variable on which to perform the analyses}
}
\value{

}
\description{

}
\examples{

dat <- data.frame(
group = rep(c("0", "30", "90"), each = 50),
my_grouping = rep(c(1:50), 3),
my_counts = rnorm(150)
)
res <- pairwise_activation(dat)
res_with_visualization <- pairwise_activation(dat, heatmap = TRUE, classifications_df = cat, classifications_cat = "parents")


}
