% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_density.R
\name{plotting_all_coordinates_2d}
\alias{plotting_all_coordinates_2d}
\title{Plot all combinations of xyz coordinates in 2D}
\usage{
plotting_all_coordinates_2d(cells_df, var_to_colour = NULL, cols = NULL)
}
\arguments{
\item{cells_df}{dataframe with xyz coordinates ("xPos", "yPos", "zPos") of the brain area of interest.
Each row corresponds to a different cell.}

\item{var_to_colour}{variable to use for colouring xyz coordinates. Default to NULL}

\item{cols}{colours to use. Must have length equal to the levels of var_to_colour. Default to NULL}
}
\value{

}
\description{
Wrapper around ggplot2 to plot xyz coordinates. Plots xy, yz, and xz coordinates and
merges the different plots together. It returns a list with 3 elements, each of which is a ggplot object.
These can be merged with other functions, such as ggpubr
}
\examples{
x <- data.frame(
sample_id = rep(c(1:3), each = 100),
xPos = c(rnorm(100), rpois(100, 2), sample(100),
yPos = c(rnorm(100), rpois(100, 2), sample(100),
zPos = c(rnorm(100), rpois(100, 2), sample(100),
group = rep(c("a", "b", "c"), each = 100)
)
my_cols <- c("black", "green", "red")
g <- plotting_all_coordinates_2d(x, "group", my_cols)
ggpubr::ggarrange(g[[1]], g[[2]], g[[3]], nrow = 3, labels = fig_labels, common.legend = TRUE, legend = "top")
}
