% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_density.R
\name{plotting_coordinates_2d}
\alias{plotting_coordinates_2d}
\title{Plot combinations of xyz coordinates in 2D}
\usage{
plotting_coordinates_2d(data_main, var_to_colour, cols, x_coord, y_coord)
}
\arguments{
\item{cells_df}{dataframe with xyz coordinates ("xPos", "yPos", "zPos") of the brain area of interest.
Each row corresponds to a different cell that belongs to a sample ("sample_id")}
}
\value{

}
\description{
Wrapper around ggplot2 to plot xyz coordinates. Used later in plotting_all_coordinates()
}
