% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_most_active.R
\name{prepare_aba_weights}
\alias{prepare_aba_weights}
\title{Prepare weights for simulation}
\usage{
prepare_aba_weights(region_df, high_prob = 0.95)
}
\arguments{
\item{region_df}{region_based dataframe. Each row is a brain area ("my_grouping") per sample ("sample_id"), where
corrected cell count ("cells_perthousand") has been summarized. It contains a variable "batch" that identifies the unit where to
perform the calculation. If from a block design, "batch" identifies a unique set control and experimental groups (var "group"), with 1 sample each.
It can be output from summarize_per_region() or preprocess_per_region().}

\item{high_prob}{number between 0 and 1 indication the threshold for being a highly active region. 0.95 corresponds to the top 5%.}
}
\value{

}
\description{
Wrapper around dplyr functions to to prepared weights for simulation study, by considering aba expression
levels as well as increase in protein expression due to the experimental manipulation.
}
\examples{
x <- data.frame(
batch = rep(c(1,1,2,2), each = 5),
group = rep(c("control", "exp", "exp", "control"), each = 5),
sample_id = rep(c("a", "b", "c", "d"), each = 5),
my_grouping = rep(c("CA1", "CA2", "CA3", "DG", "BLA"), 4),
intensity_ave = sample(10000, 20, replace = TRUE),
cells_perthousand = abs(rnorm(20))
)

find_most_active(x)
}
