% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{samples_files_to_df}
\alias{samples_files_to_df}
\title{Merges in one data frame all the separate samples' files}
\usage{
samples_files_to_df(
  path,
  extension = c("_clean_cells.RDS", "_removed_counts_summary.RDS")
)
}
\arguments{
\item{path}{String for path of location files. It does not matter if other files are present
in this folder. Only files with the specified extension will be added.}

\item{extension}{Final part of the string of the files. Same as output of cleaned_counts().}
}
\value{

}
\description{
Wrapper to dplyr functions to merge together multiple files output
from cleaned_counts(). It can merge both the cleaned cell files (extension = "_clean_cells.RDS"),
as well as the summaries of the removed cells (extension = "_removed_counts_summary.RDS").
}
\examples{
For a thorough example, please see XX.
}
