% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{normalize}
\alias{normalize}
\title{Normalize cell counts with boxcox transformation}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric vector with values to normalize}
}
\value{

}
\description{
Wrapper functon to boxcox from EnvStats to be able to normalize
in one step within pipes.
}
\examples{
x <- rpois(100, 6)
y <- normalize(x)

par(mfrow=c(1,2))
hist(x)
hist(y)
par(mfrow=c(1,1))
}
